pyproject.toml
[project]
name = "m2t"
version = "0.1.0"
description = "Convert .dm3 files into .tif files easily."
authors = [
    {name = "Trond Hauklien",email = "36609741+trondhauklien@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hyperspy>=2.4.0",
    "numba>=0.63.1",
    "pillow>=12.1.0",
    "tifffile>=2026.1.28",
    "typer>=0.21.1",
]

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
m2t = "m2t.main:app"
